rename files to support file systems which are not case-sensitive
